+++ tests/test_cell_level_utils.py
@@ -1,6 +1,5 @@
 import numpy as np
 import pytest
-
 from ccramic.app.utils.cell_level_utils import *
 from ccramic.app.parsers.cell_level_parsers import *
 from ccramic.app.inputs.cell_level_inputs import *
@@ -37,7 +36,7 @@ def test_basic_mask_boundary_converter(get_current_dir):
 
 def test_umap_from_quantification_dict(get_current_dir):
     measurements_dict = {"uploads": [os.path.join(get_current_dir, "cell_measurements.csv")]}
-    validated_measurements, cols = parse_and_validate_measurements_csv(measurements_dict)
+    validated_measurements, cols, err = parse_and_validate_measurements_csv(measurements_dict)
     returned_umap = return_umap_dataframe_from_quantification_dict(validated_measurements)
     assert isinstance(returned_umap, tuple)
     assert isinstance(returned_umap[0], dash_extensions.enrich.Serverside)
@@ -136,3 +135,45 @@ def test_convert_basic_array_to_hovertemplate():
     template = process_mask_array_for_hovertemplate(array)
     assert template.shape[2] == 1
     assert np.unique(template) == ['None']
+
+def test_get_cell_ids_in_svgpath(get_current_dir):
+    mask = np.array(Image.open(os.path.join(get_current_dir, "mask.tiff")))
+    svgpath = 'M670.7797603577856,478.9708311618908L675.5333177884905,487.2270098573258L676.0336922548805,' \
+              '492.2307545212258L671.2801348241755,500.73712044985575L669.7790114250056,' \
+              '501.98805661583077L668.0277007926405,501.4876821494408L665.7760156938856,' \
+              '499.2359970506858L663.5243305951306,497.9850608847108L662.2733944291556,' \
+              '496.23375025234577L661.7730199627656,492.9813162208108L661.7730199627656,' \
+              '491.2300055884458L662.7737688955456,490.47944388886077L665.0254539943006,' \
+              '490.47944388886077L665.7760156938856,486.4764481577408L665.2756412274956,' \
+              '484.72513752537577L664.7752667611055,482.7236396598158L666.0262029270806,' \
+              '477.2195205295258L667.2771390930556,480.7221417942558L667.5273263262505,' \
+              '481.4727034938408L668.2778880258355,479.9715800946708L668.5280752590305,479.9715800946708Z'
+    cells_included_1 = get_cells_in_svg_boundary_by_mask_percentage(mask_array=mask, svgpath=svgpath)
+    assert len(cells_included_1) == 2
+    assert list(cells_included_1.keys()) == [403, 452]
+    cells_included_2 = get_cells_in_svg_boundary_by_mask_percentage(mask_array=mask, svgpath=svgpath, threshold=0.97)
+    assert len(cells_included_2) == 1
+    assert list(cells_included_2.keys()) == [452]
+
+
+def test_basic_cell_annotation_col_pop_from_masking(get_current_dir):
+    mask = np.array(Image.open(os.path.join(get_current_dir, "mask.tiff")))
+    svgpath = 'M670.7797603577856,478.9708311618908L675.5333177884905,487.2270098573258L676.0336922548805,' \
+              '492.2307545212258L671.2801348241755,500.73712044985575L669.7790114250056,' \
+              '501.98805661583077L668.0277007926405,501.4876821494408L665.7760156938856,' \
+              '499.2359970506858L663.5243305951306,497.9850608847108L662.2733944291556,' \
+              '496.23375025234577L661.7730199627656,492.9813162208108L661.7730199627656,' \
+              '491.2300055884458L662.7737688955456,490.47944388886077L665.0254539943006,' \
+              '490.47944388886077L665.7760156938856,486.4764481577408L665.2756412274956,' \
+              '484.72513752537577L664.7752667611055,482.7236396598158L666.0262029270806,' \
+              '477.2195205295258L667.2771390930556,480.7221417942558L667.5273263262505,' \
+              '481.4727034938408L668.2778880258355,479.9715800946708L668.5280752590305,479.9715800946708Z'
+    cells_included = get_cells_in_svg_boundary_by_mask_percentage(mask_array=mask, svgpath=svgpath)
+    measurements = pd.read_csv(os.path.join(get_current_dir, "measurements_for_query.csv"))
+    assert "ccramic_cell_annotation" not in measurements.columns
+    measurements = populate_cell_annotation_column_from_cell_id_list(measurements, cell_list=list(cells_included.keys()),
+                                                                     cell_type="new_cell_type")
+    assert "ccramic_cell_annotation" in measurements.columns
+    assert len(measurements[measurements["ccramic_cell_annotation"] == "new_cell_type"]) == 2
+    assert list(measurements[measurements["cell_id"] == 1]["ccramic_cell_annotation"]) == ["None"]
+    assert list(measurements[measurements["cell_id"] == 403]["ccramic_cell_annotation"]) == ["new_cell_type"]