+++ CHANGELOG.md
@@ -11,11 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Auto sizing of the canvas now happens automatically when new ROIs are added
+- Auto sizing of the canvas now happens automatically on an ROI dimension switch
 
 ### Added
 
-- Ability to add the ccramic cell type annotation to a quantification dictiontary from a
+- Ability to add the ccramic cell type annotation to a quantification data frame from a
 region annotation: supports zoom, rectangles, and open form (non-convex) shapes
 - Added slider to change the size of the annotations (scalebar value and legend text)
 - Add toggle for custom hovertext over a mask, adding the mask/cell ID to the hover #51
@@ -24,6 +24,10 @@ region annotation: supports zoom, rectangles, and open form (non-convex) shapes
 
 - Edits to callbacks on channel switching to avoid running callbacks when the channel
 blend parameters haven't changed (happens when switching between channels in the mod menu)
+
+### Changed
+
+- More explicit warning on quantification upload if the dimensions are not compatible
 
 ## [0.5.0] - 2023-08-10